fix for cache server separated from client code; importing litefs from server as well
