+++ app/routes/admin+/cache_.sqlite.$cacheKey.ts
@@ -1,8 +1,8 @@
 import { invariantResponse } from '@epic-web/invariant'
 import { json, type LoaderFunctionArgs } from '@remix-run/node'
-import { getAllInstances, getInstanceInfo } from 'litefs-js'
 import { ensureInstance } from 'litefs-js/remix.js'
 import { cache } from '#app/utils/cache.server.ts'
+import { getAllInstances, getInstanceInfo } from '#app/utils/litefs.server.ts'
 import { requireUserWithRole } from '#app/utils/permissions.server.ts'
 
 export async function loader({ request, params }: LoaderFunctionArgs) {
